[17.0][ADD] ``account_banking_international_credit_transfer``
Add a new module to manage international credit transfers.
This adds the possibility to add an Intermediary bank on a `res.partner.bank`. When generating the pain file for an International Credit Tranfer payment method, the intermediary bank account info will be added in CstmrCdtTrfInitn > PmtInf > CdtTrfTxInf > IntrmyAgt1 > FinInstnId
For now the module only supports pain.001.001.03

A lot of the code is very close from the module ``account_banking_sepa_credit_transfer``: since a lot of copy/paste have been done, we've kept the Copyright of the original module